Claim Page refactored (almost)
Deleted the entire claim page js page. Now using htmx for the page. 375 lines of js is now just htmx and views :D
