+++ scheduler/static/css/search.css
@@ -2,6 +2,7 @@
     height: 150px;
     max-height: 150px;
     overflow-y: scroll;
+    overflow-x: none;
     user-select: none;
 }
