--- scheduler/templates/forms/form_as_p.html
@@ -1,1 +1,0 @@
-{{ form.as_p }}
