+++ scheduler/templates/pages/claim/partials/course_search.html
@@ -1,0 +1,65 @@
+<div class="row" hx-include="[name='term'], [name='department'], [name='subject'], [name='course_query']">
+    <div class="col">
+        <div class="p-2">
+            <label for="id_term">Term:</label>
+            <select hx-get="{% url 'get_course_search' %}" hx-target="#course_search" name="term" class="form-select" id="term">
+                {% for term in terms %}
+                    {% if term == selected_term %}
+                        <option value="{{ term.pk }}" selected>{{ term }}</option>
+                    {% else %}
+                        <option value="{{ term.pk }}">{{ term }}</option>
+                    {% endif %}
+                {% endfor %}
+            </select>
+        </div>
+        <div class="p-2">
+            <label for="id_department">Department:</label>
+            <select hx-get="{% url 'get_course_search' %}" hx-target="#course_search" name="department" class="form-select" id="department">
+                <option value="any" value 
+                {% if not selected_department %}selected{% endif %}
+                >Any</option>
+                <optgroup id="departmentOptions">
+                {% for department in departments %}
+                    <option value="{{ department.pk }}"
+                    {% if department == selected_department %}selected{%endif%}
+                    >{{ department }}</option>
+                {% endfor %}
+                </optgroup>
+            </select>
+        </div>
+
+        <div class="p-2">
+            <label for="id_subject">Subject:</label>
+            <select hx-get="{% url 'get_course_search' %}" hx-target="#course_search" name="subject" class="form-select" id="subject">
+                <option value="any" value 
+                {% if not selected_subject %}selected{% endif %}
+                >Any</option>
+                <optgroup id="subjectOptions">
+                    {% for subject in subjects %}
+                    <option value="{{ subject.pk }}"
+                    {% if subject == selected_subject %}selected{%endif%}
+                    >
+                        {{ subject }}
+                    </option>
+                    {% endfor %}
+                </optgroup>
+            </select>
+        </div>
+    </div>
+    <div class="col">    
+        <div class="p-2">
+            <label>Course:</label>
+            <input
+                hx-get="{% url 'get_course_options' offset=0%}"
+                hx-target="#course-options" 
+                hx-trigger="keyup changed delay:250ms, search" 
+                name="course_query" type="text" class="form-control" id="course-text" value="{{ course_query }}">
+            </input>
+            <div class="live-search-options" id="search-container">
+                <div class="container" id="course-options">
+                    {% include 'course_options.html' with offset=0 next_offset=courses|length %}
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
